requirements.txt
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bert-extractive-summarizer==0.6.1
bleach==3.2.3
blis==0.7.4
breadability==0.1.20
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
contextvars==2.4
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.21
dataclasses
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
docopt==0.6.2
eazymind==0.0.4
entrypoints==0.3
enum34==1.1.10
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
gensim==3.8.1
google-auth==1.4.2
google-auth-oauthlib==0.4.2
google-colab==1.0.0
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.8
immutables==0.15
importlib-metadata==3.4.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
jieba3k==0.35.1
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Applications==1.0.8
keras-attention==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
multiprocess==0.70.11.1
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==1.11
newspaper3k==0.2.8
nltk==3.2.4
notebook==5.2.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
pandas==0.24.2
pandas-ml==0.6.1
pandocfilters==1.4.3
parso==0.8.1
pathos==0.2.7
pathy==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
portpicker==1.2.0
pox==0.2.9
ppft==1.6.6.3
preshed==3.0.5
prettytable==2.0.0
prompt-toolkit==1.0.18
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
pysummarization==1.1.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
pyzmq==22.0.0
regex==2020.11.13
requests==2.21.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rouge==1.0.0
rsa==4.7
sacremoses==0.0.43
scikit-learn==0.20.4
scikit-plot==0.3.7
scipy==1.5.4
sgmllib3k==1.0.0
simplegeneric==0.8.1
six==1.12.0
sklearn
smart-open==3.0.0
soupsieve==2.1
spacy==3.0.3
spacy-legacy==3.0.1
srsly==2.4.0
sumy==0.7.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
textsummarizer==1.0.0
thinc==8.0.1
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tokenizers==0.10.1
tornado==4.5.3
tqdm==4.56.0
traitlets==4.3.3
transformers==4.3.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.24.3
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
